advanced-api-project/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
